requirements.txt
aiogram==3.0.0b7
SQLAlchemy>=1.4
asyncpg
python-dotenv